Renamed Palm model references to Gemini
- object names
- directory names
