+++ app/models/gemini/prompts/next_sentence.ts
@@ -20,10 +20,10 @@
 import {NextSentencePromptParams} from '@core/shared/interfaces';
 import {NextSentenceExample, WordcraftContext} from '../../../context';
 import {OperationType} from '@core/shared/types';
-import {PalmModel} from '..';
+import { GeminiModel } from '..';
 import {parseSentences} from '@lib/parse_sentences';
 
-export function makePromptHandler(model: PalmModel, context: WordcraftContext) {
+export function makePromptHandler(model: GeminiModel, context: WordcraftContext) {
   function generatePrompt(
     textBeforeBlank: string,
     textAfterBlank: string,
