+++ app/models/gemini/prompts/rewrite_selection.ts
@@ -21,13 +21,13 @@ import {shuffle} from '@lib/utils';
 import {RewriteSelectionPromptParams} from '@core/shared/interfaces';
 import {RewriteSelectionExample, WordcraftContext} from '../../../context';
 import {OperationType} from '@core/shared/types';
-import {PalmModel} from '..';
+import { GeminiModel } from '..';
 
 function capitalize(str: string) {
   return str.charAt(0).toUpperCase() + str.slice(1);
 }
 
-export function makePromptHandler(model: PalmModel, context: WordcraftContext) {
+export function makePromptHandler(model: GeminiModel, context: WordcraftContext) {
   function insertBlank(pre: string, post: string) {
     return `${pre}${model.getBlank()}${post}`;
   }
