+++ app/models/gemini/prompts/rewrite_sentence.ts
@@ -21,9 +21,9 @@ import {shuffle} from '@lib/utils';
 import {RewriteSentencePromptParams} from '@core/shared/interfaces';
 import {RewriteSentenceExample, WordcraftContext} from '../../../context';
 import {OperationType} from '@core/shared/types';
-import {PalmModel} from '..';
+import { GeminiModel } from '..';
 
-export function makePromptHandler(model: PalmModel, context: WordcraftContext) {
+export function makePromptHandler(model: GeminiModel, context: WordcraftContext) {
   function insertBlank(pre: string, post: string) {
     return `${pre}${model.getBlank()}${post}`;
   }
